move TupleSplit to schema-type
